Move throw tests to own file
